Updated inferredEnd logic
`@to` values with the same numeric content as `@from` will now only appear if both values are non-numeric. That is:

If `@from`='16' and computed `@to`='16r', `@to` will not be included.

If `@from`='16r' and computed `@to`='16', `@to` will not be included.

If `@from`='16r' and computed `@to`='16v', `@to` will be included.